rearranged test.java files into single package
